// src/app/model/cliente.ts
import { TipoCarteiraHabilitacao } from './tipoCarteiraHabilitacao';

export class Cliente {

  status: number;
  id: number;
  nomeCompleto: string;
  caminhoFoto: string;
  contentType: string;
  dataNascimento: Date;
  sexo: number;
  endereco: string;
  bairro: string;
  cidade: string;
  estado: string;
  cep: string;
  telefone: string;
  email: string;
  rg: string;
  cpf: string;
  representante: string;
  financiamento: string;
  entrevistadopor: string;
  profissao: string;
  tipoVisto: number;
  vencimentoVisto: Date;
  numeroPassaporte: number;
  nacionalidade: string;
  cidadeNatal: string;
  estadoCivil: number;
  tipoSanguineo: number;
  altura: number;
  manequim: number;
  peso: number;
  calcado: number;
  grauEscolaridade: number;
  tipoCarteiraHabilitacaoList: TipoCarteiraHabilitacao[];
  conversacao: number;
  compreensao: number;
  leitura: number;
  escrita: number;
  conheceuPorOnde: string;
  conhecidosNaLiberdadeRh: boolean;
  shakaiHoken: boolean;
  seguroDesemprego: boolean;
  outraLinguas: boolean;
  tempoJapao: boolean;
  preferenciaRegiao: boolean;
  mudanca: boolean;
  tatoo: boolean;
  bebidasAlcoolicas: boolean;
  divida: boolean;
  problemaPolicialJapao: boolean;
  destroCanhoto: boolean;
  dautonico: boolean;
  alergia: boolean;
  doencaGrave: boolean;
  deficienciaFisica: boolean;
  tomaRemedio: boolean;
  problemaDeVista: boolean;
  problemaColuna: boolean;
  fumante: boolean;
  gravida: boolean;
  diabete: boolean;
  password: string;
  empresa: boolean;
  outrasLinguasRendered: string;
  preferenciaRegiaoRendered: string;
  tatooRendered: string;
  alergiaRendered: string;
  ativo: boolean;
  idade1: string;
  zairyuCard: string;
  localidade: number;

}